+++ social_core/backends/kakao.py
@@ -21,7 +21,9 @@ class KakaoOAuth2(BaseOAuth2):
 
     def get_user_details(self, response):
         """Return user details from Kakao account"""
-        kaccount_email = response.get('kaccount_email', '')
+
+        kakao_account = response.get('kakao_account', '')
+        kaccount_email = kakao_account.get('email', '')
         properties = response.get('properties', '')
         nickname = properties.get('nickname') if properties else ''
         return {
@@ -35,7 +37,7 @@ class KakaoOAuth2(BaseOAuth2):
     def user_data(self, access_token, *args, **kwargs):
         """Loads user data from service"""
         return self.get_json(
-            'https://kapi.kakao.com/v1/user/me',
+            'https://kapi.kakao.com/v2/user/me',
             headers={
                 'Authorization': 'Bearer {0}'.format(access_token),
                 'Content_Type': 'application/x-www-form-urlencoded;charset=utf-8',